fix: handle strings with unescaped forward slashes
Problem:
I was sending strings to regexp-tree instead of regexes.
regexp-tree strings need to have '/.../' so I was wrapping regexes-as-strings with forward slashes.
If the string already contained an unescaped forward slash, then it would no longer be a valid regex, and parsing it would fail.
Example: ^abc/ --> /^abc//
If safe-regex can't parse a regex, it returns false.

Solution:
Send RegExp objects to regexp-tree.

Fixes: #20
https://github.com/davisjam/safe-regex/issues/20